pyproject.toml
[project]
name = "xlw_exp"
version = "0.1.0"
description = "Federated Learning with FAP-Float(S) and EAPH-LoRA for DomainNet"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=1.12.0",
    "torchvision>=0.13.0",
    "numpy>=1.24.0",
    "pillow>=9.0.0",
    "pyyaml>=6.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[dependency-groups]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["data", "models", "core", "utils"]